// src/lib/api/generated/model/customerListItemDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Agendly.Api
 * OpenAPI spec version: 1.0
 */

export interface CustomerListItemDto {
  id: string;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  completedCount: number;
  noShowCount: number;
  totalSpent: number;
  /** @nullable */
  lastVisitAt?: string | null;
}